Fix deadlock when client send to serverboards an unknown method.
It was a deadlock as the answer for unknown method required to
use the server that initially try to process the method call.

Now it uses cast, so it queues the writef for later.

It may happen that its not writen, but clients thinks it is,
but this does not ensure neither that the network will send
the message properly, so not loosing anything.