Fix method caller when mc [:nok, :ok]
There was an edge case when a method caller has other method callers inside,
and the first fails and the second succeeds. It was signaling
incorrecly that the first fails, and giving the work to the second,
causing sending two responses to the reply cahnnel, one
with the answer and another with the :method_unknown.