Working update service with components
Updating a service with a component key will update those components as well.

Fixes that before was trying to use the eventsourcing mechanisms to retrigger an event, and that was a recipe for failures. Now the command events do the full job, invoking if necesary real component updters.

This required as well to restructure how to components receive the event sourcing events and dispatch them, just calling a function that does the real state change.

Also the component change, although at :service_update event its processed at component, just listening for the same event.