Working subscriptions, improved preload component data
This required creating a wrapper around redux-react connect that makes the subscriptions at enter and unsubscribes at unmount. Also dispatches the requests for initial data, as it will be stale after unsubscribing.

Subscription uses a refcount to do not send unneeded subscription requests, and do not unsubscribe if somebody still listening.