Added `app` alias for root of serverboards app.
Modified all uses of relative imports to use app when the intended relative use
was to go to root and then to another subsystem.

Due to the separation of modules by concern was quite common from one
module from containers to refer to one components and from there to an action,
for example.