Allow contexted subscriptions and use on rules
Now its possible to subscribe to an event given a context in
the form of `event[context]`. The emmiter should send both
with and without context, and if the client is subscribed,
then the emited event is just `event`.

It is this way so that the client can subscribe to narrow
events (for example rule changes at a serverboard only), but
to ease the implementation as it should only check for the
`rules.update` event.