WIP Improvements on cmd fail detection
The gist is a monitor and trapexit so that we catch when the port fails
and then stop the Cmd. This will cascade to the Action runner and log
properly a failed action.

Other callers have to manage failed plugins in their own ways, but
always based on exited Cmd on error.