Can add widgets, use widgets at serverboard
Also added support for exporting specific componetns from the core to
the plugins, just now MarkdownPreview, as they can be heavy components
to include on each plugin and can reuse the one in the core. In the
future may have as well custom elements that we want to be used on
plugins.

Added markdown plugin.