Added watched properties to event.subscribe_connect
This allows to only recalculate subscriptions if the given property
changes.

This is used for very specific situations, for exampel the board at
serverboard overview, which can change to a new board directly,
just changing props, and React will just rewrite leaf elements,
leaving the main component intact and not regenerating it, only changing
props. Without watching for the serverboard name prop, we can only or
gather all the facts and subscriptions for every prop change on all the
event.subscribed components. With it we only do that when the
serverboard name changes.