Triggers now have start and stop functions
Now they can not block, as the executable may be reused for several
triggers.

They canges are update to latest python bindings taht allow timers, and
use in tests.

Also keep more state in the trigger itself and always pass a trigger id
that should be sent back on trigger calls.

The trigger stop function is called with the result of the start call or
if none, with the id of the trigger.