Plugin runner call can be with a method struct to filter params
This eases the use of the calls, as can be with any param and the method
definition is then used as filter, setting even defaults. Also
simplifies the definition at manifests as can be simple method name if
it can receive any param (and do not need configuration) or the full
struct.