Async widget adding to the factory and use
Before required strongly it was synchronous (at least in the
proper order), now it uses promises to accept the possiblity that
the HTML/JS we load does not add the constructor straigth away,
and have some delay.

But if in some seconds is not ready it fails.

Also some message on widget load failures.