Reworked rules, more natural architecture
Before it was a mix of the triggers and the rule, with some complex
management.

Now each rule is a process, properly supervised. It does all that the
rule requires, taht before was in trigger.

Also moved the evensourcing functions to rules, so now all db changes
are performed at rules.