Fix race condition for registry and use of rule id
The registration of the name was after the trigger start was called,
and in the client side a trigger may be thrown by the start function.

Now we register before any call, and if there is a trigger event, it
will be properly queued.

Also control better the situation where the rule does not exist anymore
(or ever) and the client triggers an event.