Can open `external url` components in service details
A external URL is a UI component that can create a url from a pattern
from the config of the service.

It allows easy access to external resources that can work on your
service, like the URL for the service itself, or web apps as google
pagespeed or ssl labs.