Added RPC call plugin.is_running, and use to relaunch plugin.start
This means that plugins that use plugin.start can be assured that the
call will succeed, even if its in another instance of the plugin.

This is usefull for short lived widget helper, for example, as they will
succed. But ht eplugin has to be restart aware, and be able to set up
initial state properly by its own.