Added utils/cache to ease get state, added services redux state
The new cache was needed as asking for some dta from server may be
costly and is not required foreach screen. Previously it was asuming it
had it from a aprent component, or not having it at all, when some times
was already loaded.

With teh cache services if its on cache, it will do nothing, and if its
not ready it will load this data. The cache functions return a proper
promise, and update the store. It uses the store to store the data.

Also added some order to the services, so that now it uses a proper
store key of services.

Exported some extra symbols for plugins.