Added new store.connect
This new connect will replace the old event.connect, and provides a
saner API to allow use of futures, event subscription and so on.

This allows to sidestep the use of the redux store for individual views
that are too specific to use the general store.

Use of this on the service details view.