Pass layout information to widgets. Load CSS at screens and widgets
Its passed as an object in the extra config data, and set into the DOM
as data-width and data-height into the widget card.

The data attributes are useful to use CSS to change slightly the layout
depending on the card size.

Now it can load extra CSS for the widgets and screens. Proper CSS
scoping is required.