Fix bug in FF where object has a watch property always
This is the watch object set, and was making it all fall apart.

This property signalled to connect that it had to watch for some props
changes, but did not convert to a proper map, producing an exception and
making react just fail.

Now we check that the options.watch != Object.watch.