[WIP] Can load translations from json file
It gathers the data into a po file, to be translated with external
tools, and updates a JSON file loaded at runtime after login which
will set up the translations and use them.

Then system takes care of using the translations, but for some semantic
ui elements it is still failing on auto update; it requires reload of
the element.

Added initial es translations and script to generate the JSON.