Added support for shallow triggers
Shallow triggers are triggers that do not start a funtion to check the
status of some condition, but rely on external helper, as an init
or cron to properly do the checkings.

One way to make this work is to use a cron that periodically checks some
condition, then checks for rules cheking for a specific trigger type,
and compare and triggers the status changes as required.