Fix out of order reply on python bindings
Due to more asynchronicity, a bug appeared in the python bindings.

It was now possible to do several UI calls to the plugin parallelized,
and so the plugin may call the server for some info, parallelized, and
get the answers in the opossite order of ask, for example this order: Q1
Q2 A1 A2, so whicle Q2 is waiting for A2, get A1.

The solution is to store it for later, and keep waiting. When Q2 gets
the answer, and returns finally to the waiting loop for Q1, it first
checks if already answered, and if so uses the A1 properly. If not,
keeps waiting as usual.