Triggers are async to prevent deadlock
Rule triggers were causing a deadlock.

When the trigger starts, it calls a function in the plugin code that may
trigger a new state directly at this function. If that happens, as rule
process is already waiting the start to finish, it deadlocks.

As a solution the trigger itself is now asynchronous. It always return
:ok, so it signal not that it was done, but aknowledged.

Then, when the init function finishes, the trigger will be next in the
queu and propiate action will be called.

Before this bug was not triggered as the calls were serialized, and the
trigger call was not delivered until init was over (I think).