Do not give the option of keep logged in
It always keeps logged in. If want to log out, use log out.

Session expires after 24h of no use. May reduce time, but no need to
tell the system to keep logged in.

The old behaviour (only reload kept logged in, other tabs not) was not
intuitive nor useful. When accessing serverboards from other sources
(oauth flow), it kept requireing the password, when intuition says it
should not.