WIP: Works 3-way communication
Plugin can make requests to core, core to user, user to both... all
expected connections.

This allows to execute a plugin and it will require the core whatever it
need, and the user can do requests to the plugin or to the core.

For that it uses the '.' and a whitelist to check where to ask when
comming from user.

It is generic, but requires manual connecting functions.