Fixed communication between process and core and cli
Uses non blocking PIPE to communicate with the command.

There was some cases where communication was not properly retried which
was needed as there might be several cmd<->core communications. They
were forced forcing a communication loop. Now when there is
communication it is registered, and if necesary repeat the loop straight
away. This increases the interactivility a great deal.