Can add nodes to the rule
There is a new node type, add, which is used to mark that it can add.

This was required to allow navigation.

Rules are preprocessed, adding add nodes, and should be removed before
inserting in the database (TODO).

Navigation works, and adding a node mutates the node to the new type,
and then adds an extra `add` node.