Improved show logs on plugin test error
Added `until` to logs.history to show from last message until some id.

The s10s plugins-test grabs the message id of a new message and then
uses this id as until. In case of fail, shows a basic log printout of
all messages from the start of the test till the end.

This is not a full picture of what happeneed, as it may even have non
related messages from other background running tasks, but gives some
ideas.