Improved return messages
Returns False instead of -1 at file closed read/write as for valid block
test -1 may fail, and both "" and False are falsy.

This simplifies some code where the only important point is if there was
read/write data and at the same time allows to know the reason when
using nonblock mode.