Pass the ready/cloed fd to the event
This is required as subscribing to an event with filter does not deliver
the fitlered event, but the plain one. If the user filters by fd
descriptor, and uses the same handeler for several files, there will be
no way to know which descriptor created the call, as all of them will go
to the handler.

Internally each handler should check that it received the proper
descriptor it was watching.