Can template maps. Leave unknowns as is.
The biggest change is that before tempaltes that did not know a variable
were printing an error (with actually not valid data). Now it returns
the same placeholder. This can be used as an error mark, but also to
iteratively complete the template.

This is used for rule templates, were some data is known at the rule
execution, but some only at execution time.