Reworked how the top menu is changed, to avoid redraws
Redraws at initial state were provoking a leak of React components, as
the destructor was not properly called.

Now it works in a two phase way, which is less efficient, but more sure.

It uses the Top component to store the state of the menus, and a
callback to set the handlers. Until the handlers are set, there is no
main section, once set, the section is shown.

This seciton may in turn set the top menu without provoking a redraw of
the shown section.