Initial values for url and email address at ini
Modified precendence order at config value get, and added support for
plugin settings daa at .ini file.

The config for plugins requires chane the / for a - in the section name,
as / is not a valid ini section name character.

It could have been added manually to the parser, but at using a diferent
separator was choosen as a better option.