Fix restart plugin at exit at py bindings
It should restart the plugin (if allowed) when the plugin exited (quite
probably timeout) and another method is called. It was working, but
changes in how the exit is signaled made it not work on cloud component.

Now timeouts will not affect how the cloud component is used; should
just work always, restarting the vmc proviers as needed.