Fix init timeouts and reloads
Was double resetting init processes, which was doubling time between
retries. Now takes  care of let the current timer expire y detected a
reason to retry. It happens when  the process ends properly and sets the
timer, but then its detected that the  process is down and sends another
timer with  the double time.