Fix plugindata test was using other ecto connection
For unknown reasons somtimes the plugindata_test was using another ecto
connection that was setting the data on the database, but inside a
transaction,
which was not visible for the reader process.

Now it forces stop and restart ecto before the tests start, so it
snsures it is
in the same transaction.