Do not crash if invalid insert into cache
This happened when the data takes time to get calculated and another
task asks for the same data. Then the second timeout, removes the
waiting pids, including original caller. When the original caller has
the result, tries to set it and then it fails as there is no list of
tasks waiting.

The fix sets the default list to [], so there is no crash, but also adds
the timeout option to make it independant of TTL which may be
differently lived.

Timeout only applies for succesive calls, not for the initial caller.