Generic forms items may wait for an event
This allows to have a background component change a value in the form.
This is
required for example for Oauth authentication which is fully async and
uses
several tabs to complete.

The other tab sends an event with the proper value and then this forms
gets the data.

The event is a free text which will be added a form_id to the end. This
form_id
is received by the "vars" at richtext and buttons.