WIP New SectionMenu workings
Uses some react hacks to pass the menu component and props through a 
SectionMenu component.

Does not use new React Context as it is ill fitted for this pourpose. 
Could have
used old context, but as topmenu is unique per Serverboards, it can be 
solved
using singletons and forcing updates.