Fix checkbox false at generic form transforms to true
When a generic form with a checkbox is not marked, at reload of the form 
it did
not see the proper value, and set it to true. So it was difficult to 
keep it as
false.

Now checks for proper undefined value and then sets the false if 
appropiate, or
default values.