Fix timer runaway at init fail/restart
On some conditions the init system will keep creating timers to try to 
make the
init work, but not cancel previous ones, which leads to a parallel 
triggering of
the restart and if fail, then even more. 

This was detected with millions of entries at the logs with restarts at 
exactly
the same time.